+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.0] - 2019-03-05
+
+### Changed
+
+- Support using with production builds of the current version of Preact 10
+  (10.0.0-alpha0).
+
 ## [1.6.0] - 2019-02-22
 
 ### Added